+++ rio/components/pointer_event_listener.py
@@ -138,9 +138,6 @@ class PointerEventListener(FundamentalComponent):
     `on_double_press`: Similar to `on_press`, but triggered when the mouse
         button is double-pressed.
 
-    `double_press_delay`: The maximum time in seconds between two presses that
-        should be considered a double press.
-
     `on_pointer_down`: Triggered when a pointer button is pressed down while
         the pointer is placed over the child component.
 
@@ -171,9 +168,14 @@ class PointerEventListener(FundamentalComponent):
     _: dataclasses.KW_ONLY
     on_press: rio.EventHandler[PointerEvent] = None
     on_double_press: rio.EventHandler[PointerEvent] = None
-    double_press_delay: float = 0.3
-    on_pointer_down: rio.EventHandler[PointerEvent] = None
-    on_pointer_up: rio.EventHandler[PointerEvent] = None
+    on_pointer_down: (
+        rio.EventHandler[PointerEvent]
+        | dict[MouseButton, rio.EventHandler[PointerEvent]]
+    ) = None
+    on_pointer_up: (
+        rio.EventHandler[PointerEvent]
+        | dict[MouseButton, rio.EventHandler[PointerEvent]]
+    ) = None
     on_pointer_move: rio.EventHandler[PointerMoveEvent] = None
     on_pointer_enter: rio.EventHandler[PointerEvent] = None
     on_pointer_leave: rio.EventHandler[PointerEvent] = None
@@ -185,7 +187,6 @@ class PointerEventListener(FundamentalComponent):
         return {
             "reportPress": _list_buttons_to_report(self.on_press),
             "reportDoublePress": _list_buttons_to_report(self.on_double_press),
-            "doublePressDelay": self.double_press_delay,
             "reportPointerDown": _list_buttons_to_report(self.on_pointer_down),
             "reportPointerUp": _list_buttons_to_report(self.on_pointer_up),
             "reportPointerMove": self.on_pointer_move is not None,
@@ -211,7 +212,7 @@ class PointerEventListener(FundamentalComponent):
             )
 
         elif msg_type == "doublePress":
-            await self.call_event_handler(
+            await self._call_appropriate_event_handler(
                 self.on_double_press,
                 PointerEvent._from_message(msg),
             )
@@ -294,7 +295,8 @@ PointerEventListener._unique_id_ = "PointerEventListener-builtin"
 
 
 def _list_buttons_to_report(
-    handler: rio.EventHandler[PointerEvent] | None,
+    handler: rio.EventHandler[PointerEvent]
+    | dict[MouseButton, rio.EventHandler[PointerEvent]],
 ) -> Jsonable:
     if handler is None:
         return []